Exception and Advice

+++ src/main/java/com/example/wineshop/RegionNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.wineshop;
 
 public class RegionNotFoundException extends RuntimeException {
 
-    RegionNotFoundException(Long id){
+    RegionNotFoundException(int id){
         super("Could not find region "+ id);
     }
 }
