+++ src/main/java/com/example/wineshop/TypeNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.wineshop;
 
 public class TypeNotFoundException extends RuntimeException {
 
-    TypeNotFoundException(Long id){
+    TypeNotFoundException(int id){
         super("Could not find type "+ id);
     }
 }
