+++ src/main/java/com/example/wineshop/WineNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.wineshop;
 
 class WineNotFoundException extends RuntimeException {
 
-    WineNotFoundException(Long id){
+    WineNotFoundException(int id){
         super("Could not find wine "+ id);
     }
 }
