+++ src/main/java/com/example/wineshop/WineryNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.wineshop;
 
 public class WineryNotFoundException extends RuntimeException {
 
-    WineryNotFoundException(Long id){
+    WineryNotFoundException(int id){
         super("Could not find winery "+ id);
     }
 }
